pyproject.toml
[project]
name = "traceback"
version = "0.1.0"
description = "Traceback - Enterprise Knowledge Agent Platform for Incident Response"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.1",
    "atlassian-python-api>=4.0.7",
    "beautifulsoup4>=4.14.2",
    "datasets>=4.3.0",
    "fastapi>=0.120.0",
    "google-cloud-firestore>=2.21.0",
    "google-cloud-secret-manager>=2.25.0",
    "httpx>=0.28.1",
    "langchain>=1.0.2",
    "langchain-google-genai>=3.0.0",
    "langchain-qdrant>=1.1.0",
    "langgraph>=1.0.1",
    "lxml>=6.0.2",
    "neo4j>=5.0.0",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "python-gitlab>=6.5.0",
    "ragas>=0.3.7",
    "rank-bm25>=0.2.2",
    "structlog>=25.4.0",
    "uvicorn[standard]>=0.38.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=1.3.0",
]
